+++ frontend/src/ui/pages/AboutPage/AboutPage.js
@@ -1,10 +1,9 @@
 import React, { useState } from 'react'
 import './AboutPage.scss'
-import Page1 from './Page1'
-import Page2 from './Page2'
-import Page3 from './Page3'
-import { Link } from 'react-router-dom'
+import { Link, useNavigate } from 'react-router-dom'
 import arrow from '../../assets/other/arrow.png'
+import ChatBubble from '../../components/ChatBubble/ChatBubble'
+import { Button } from '../../components'
 
 function AboutPage() {
 	const [currentPage, setCurrentPage] = useState(1)
@@ -13,24 +12,37 @@ function AboutPage() {
 		setCurrentPage(pageNumber)
 	}
 
+	const navigate = useNavigate()
+
 	const renderChatBubbles = () => {
 		switch (currentPage) {
 		case 1:
 			return (
-				<div>
-					<Page1 />
+				<div className='chat-container'>
+					<ChatBubble text={'Hey there, what\'s \nDo The Be(a)st all about?'}/>
+					<ChatBubble side='right' text={'Greetings, adventurer! \nIn Do The Beast, we turn your to-do list into \nepic quests. Create lists and our AI will prioritize \ntasks and serve you one monstrous challenge at a time.'} />
+					<ChatBubble text={'How do I \ndefeat these monsters?'}/>
+					<ChatBubble side='right' text={'It\'s simple! Your tasks represent epic battles. \nThe monsters have HP, and your tasks are your weapons. \nWhen you complete a task, it chips away at \nthe monster\'s health. Defeat monsters and earn trophies.'}/>
 				</div>
 			)
 		case 2:
 			return (
-				<div>
-					<Page2 />
+				<div className='chat-container'>
+					<ChatBubble text={'Trophies, you say?'}/>
+					<ChatBubble side='right' text={'\nIndeed! Conquer a monster, and it will add its trophy \nto your collection. Display them proudly \nin your virtual trophy room.'} />
+					<ChatBubble text={'Can I team up with others?'}/>
+					<ChatBubble side='right' text={'\nOf course! Join the DoTheBeast community, where you can \nshare your conquests, \nstrategize, and embark on group quests with \nfellow adventurers.'}/>
 				</div>
 			)
 		case 3:
 			return (
-				<div>
-					<Page3 />
+				<div className="chat-container">
+					<ChatBubble text={'This sounds like an \nepic adventure!'}/>
+					<ChatBubble side='right' text={'\nIndeed, it is! Welcome to DoTheBeast, where productivity \nbecomes a thrilling game. Your adventure awaits!'} />
+					<div className='buttons-container-about'>
+						<Button onClick={() => navigate('/login')}>LOGIN</Button>
+						<Button onClick={() => navigate('/register')}>SIGNUP</Button>
+					</div>
 				</div>
 			)
 		default:
@@ -40,10 +52,7 @@ function AboutPage() {
 
 	return (
 		<div>
-			<div>
-				{renderChatBubbles()}
-			</div>
-
+			{renderChatBubbles()}
 			<div className="navigation-dots">
 				<div className="dot" onClick={() => handlePageChange(1)}></div>
 				<div className="dot" onClick={() => handlePageChange(2)}></div>
